Rename warnings to warnings-db
